+++ src/CustomType/Data/Slice/Slice.php
@@ -6,9 +6,11 @@ use Torr\PrismicApi\CustomType\Data\Field\InputField;
 use Torr\PrismicApi\CustomType\Data\PrismicTypeInterface;
 
 /**
+ * You can extend this class to create reusable slices
+ *
  * @see https://prismic.io/docs/core-concepts/slices
  */
-final class Slice implements PrismicTypeInterface
+class Slice implements PrismicTypeInterface
 {
 	/**
 	 * @param array<string, InputField> $fields
@@ -27,7 +29,7 @@ final class Slice implements PrismicTypeInterface
 
 	/**
 	 */
-	public function toArray () : array
+	final public function toArray () : array
 	{
 		return [
 			"type" => "Slice",